add missing FLAG_SUPPORTS_REMOVE, move ViewModel file
