+++ packages/terre-android/app/src/main/java/com/openwebgal/terre/MainActivity.kt
@@ -39,6 +39,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.openwebgal.terre.ui.theme.TerreTheme
+import com.openwebgal.terre.viewmodel.TerreViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
